Rename mainImpl -> main-impl
