+++ packages/pyright-scip/src/main-impl.ts
